Mostly managing to run a fragmented alignment fully in a local hadoop job.  Currently it hangs after generating the sam file though.

+++ hbase-genomes/src/main/java/org/lcsb/lu/igcsa/BWAIndex.java
@@ -32,7 +32,7 @@ public class BWAIndex extends Configured implements Tool
     new BWAIndex().run(args);
     }
 
-  // TODO this only runs if bwa is installed on your system
+  // TODO this only runs if bwa is installed on your system - also a bit broken at the moment
   public int run(String[] args) throws Exception
     {
     this.fasta = new Path(args[0]);
@@ -46,8 +46,7 @@ public class BWAIndex extends Configured implements Tool
     // need a local directory to copy things to if it doesn't already exist
     if (fs.getUri().toASCIIString().equals("hdfs:///"))
       {
-      if (localFasta.getParentFile().exists())
-        FileUtils.cleanDirectory(localFasta.getParentFile());
+      if (localFasta.getParentFile().exists()) FileUtils.cleanDirectory(localFasta.getParentFile());
       else localFasta.getParentFile().mkdir();
       }
 
@@ -96,12 +95,12 @@ public class BWAIndex extends Configured implements Tool
     return 0;
     }
 
-  public StreamWrapper getStreamWrapper(InputStream is, String type)
+  public static StreamWrapper getStreamWrapper(InputStream is, String type)
     {
     return new StreamWrapper(is, type);
     }
 
-  private class StreamWrapper extends Thread
+  public static class StreamWrapper extends Thread
     {
     InputStream is = null;
     String type = null;
@@ -139,4 +138,94 @@ public class BWAIndex extends Configured implements Tool
       }
     }
 
+  /*
+  This doesn't work.  BWA isn't able to read the file from hdfs even when run within a mapper.
+   */
+//  public int run(String[] args) throws Exception
+//    {
+//    //args = new String[]{"/tmp/GRCh37-kiss1/merged.txt"};
+//    String input = args[0];
+//
+//    log.info(input);
+//
+//    Path inputPath = new Path(input);
+//
+//    Configuration conf = new Configuration();
+//
+//    // TODO this is a temporary measure
+//    URI uri = new URI("hdfs://localhost:9000/bwa-tools/bwa.tgz#tools");
+//
+//    // WARNING: DistributedCache is not accessible on local runner (IDE) mode.  Has to run as a hadoop job to test
+//    DistributedCache.addCacheArchive(uri, conf);
+//    DistributedCache.createSymlink(conf);
+//
+//    Job job = new Job(conf, "BWA Index");
+//    job.setJarByClass(Test.class);
+//
+//    job.setMapperClass(IndexMapper.class);
+//    job.setInputFormatClass(KeyValueTextInputFormat.class);
+//    //job.setInputFormatClass(WholeFileInputFormat.class);
+//    //    job.setMapOutputKeyClass(Text.class);
+//    //    job.setMapOutputValueClass(Text.class);
+//    FileInputFormat.addInputPath(job, inputPath);
+//
+//    //job.setReducerClass(IndexReducer.class);
+//    job.setNumReduceTasks(0);
+//    job.setOutputFormatClass(NullOutputFormat.class);
+//
+//    //FileOutputFormat.setOutputPath(job, inputPath.getParent());
+//
+//    return (job.waitForCompletion(true) ? 0 : 1);
+//
+//    }
+//
+//
+//
+//  static class IndexMapper extends Mapper<Text, Text, Text, Text>
+//    {
+//    private String bwaBinary = "tools/bwa";
+//
+//    @Override
+//    protected void setup(Context context) throws IOException, InterruptedException
+//      {
+//      if (!new File(bwaBinary).exists()) throw new RuntimeException("bwa binary does not exist in the cache.");
+//
+//      log.info("BWA BINARY FOUND");
+//
+//      super.setup(context);
+//      }
+//
+//
+//    @Override
+//    protected void map(Text key, Text value, Context context) throws IOException, InterruptedException
+//      {
+//
+//      String cmd = bwaBinary + " index -a bwtsw " + value.toString();
+//      log.info("BWA cmd: " + cmd);
+//
+//      Process p = Runtime.getRuntime().exec(cmd);
+//
+//      BWAIndex.StreamWrapper error, output;
+//      error = BWAIndex.getStreamWrapper(p.getErrorStream(), "ERROR");
+//      output = BWAIndex.getStreamWrapper(p.getInputStream(), "OUTPUT");
+//      int exitVal = 0;
+//
+//      error.start();
+//      output.start();
+//      error.join(3000);
+//      output.join(3000);
+//      exitVal = p.waitFor();
+//
+//      log.info("-- " + exitVal + " --");
+//      log.info("BWA OUTPUT: " + output.message);
+//      log.info("BWA ERROR: " + error.message);
+//
+//      if (exitVal > 0) throw new RuntimeException("BWA error: " + error.message);
+//
+//      //super.map(key, value, context);
+//      }
+//    }
+
+
   }
+
